Move styleguide installation into install task
- Updated USAGE for change in task name for install
- Output USAGE info at the end of install
