+++ lib/generators/styleguide/USAGE
@@ -2,9 +2,10 @@ Description:
     Add a /styleguide route to your app to preview CSS styles
 
 Example:
-    rails generate styleguide
+    rails generate styleguide:install
 
     This will create:
         app/controllers/styleguide_controller.rb
         app/views/styleguide/
+        app/views/styleguide/widgets/
         appropriate routes
